requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-channels-graphql-ws==0.8.0
django-filter==2.4.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
install==1.3.4
msgpack==1.0.2
multidict==5.1.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
yarl==1.6.3
zope.interface==5.4.0
